`pyrogram v2 here`
- please report all bugs.
- new admins commands added.
- new callback data added.
- background code error fixed.
- and many more help to you.

+++ Rose/plugins/ban.py
@@ -1,43 +1,36 @@
 from time import time
 import asyncio
-from pyrogram.types import Message
 from Rose.utils.extract_user import extract_user
 from Rose import BOT_ID,app
 from Rose.utils.functions import extract_user_and_reason,time_converter
-from pyrogram.types import Message
-from lang import get_command
-from Rose.utils.commands import *
-from Rose.utils.lang import *
+from pyrogram.types import (
+    ChatPermissions,
+    InlineKeyboardButton, 
+    InlineKeyboardMarkup,
+    Message)
+from Rose.utils.commands import command
+from Rose.utils.lang import language
 from Rose.utils.custom_filters import restrict_filter
 from Rose.plugins.fsub import ForceSub
-from button import *
+from button import Restrict
 
-KICK_ME = get_command("KICK_ME")
-SKICK = get_command("SKICK")
-KICK = get_command("KICK")
-BAN = get_command("BAN")
-UNBAN = get_command("UNBAN")
-SBAN = get_command("SBAN")
-
-
-@app.on_message(command(KICK_ME) )
+@app.on_message(command("kickme") )
 @language
 async def kickFunc(client, message: Message, _):
     reason = None
     if len(message.text.split()) >= 2:
         reason = message.text.split(None, 1)[1]
     try:
-        await message.chat.ban_member(message.from_user.id)
-        txt = f" Bye {message.from_user.mention}, you're right - get out."
+        await app.ban_chat_member(message.chat.id,message.from_user.id)
+        txt = f" Bye {message.from_user.mention}, you're right."
         txt += f"\n<b>Reason</b>: {reason}" if reason else ""
         await message.reply_text(txt)
-        await message.chat.unban_member(message.from_user.id)
+        await app.unban_chat_member(message.chat.id,message.from_user.id)
     except Exception as ef:
         await message.reply_text(f"{ef}")
     return
 
-
-@app.on_message(command(SKICK) & restrict_filter)
+@app.on_message(command("skick") & restrict_filter)
 @language
 async def kickFunc(client, message: Message, _):
     if len(message.text.split()) == 1 and not message.reply_to_message:
@@ -47,14 +40,10 @@ async def kickFunc(client, message: Message, _):
     except Exception:
         return   
     if not user_id:
-        await message.reply_text(_["ban2"])
-        return  
+        return await message.reply_text(_["ban2"])  
     st = await client.get_chat_member(message.chat.id, user_id)
-    if (
-                st.status == "administrator"
-                and st.status == "creator"
-        ):
-            return
+    if(st.status == "administrator" and st.status == "creator"):
+        return
     try:
         await message.chat.ban_member(user_id)
         await message.delete()
@@ -62,32 +51,22 @@ async def kickFunc(client, message: Message, _):
             await message.reply_to_message.delete()
         await message.chat.unban_member(user_id)
     except Exception as ef:
-        await message.reply_text(f"{ef}")
+        await message.reply_text(ef)
     return
 
-
-
-@app.on_message(command(KICK) & restrict_filter)
+@app.on_message(command("kick") & restrict_filter)
 @language
 async def kickFunc(client, message: Message, _):
-    FSub = await ForceSub(app, message)
-    if FSub == 400:
-        return
     user_id, reason = await extract_user_and_reason(message, sender_chat=True)
     if not user_id:
         return await message.reply_text(_["ban3"])
     if user_id == BOT_ID:
         return await message.reply_text(_["ban4"])
     st = await client.get_chat_member(message.chat.id, user_id)
-    if (
-                st.status == "administrator"
-                and st.status == "creator"
-        ):
+    if (st.status == "administrator" and st.status == "creator"):
         return await message.reply_text(_["ban7"])
     mention = (await app.get_users(user_id)).mention
-    msg = f"""
-{mention}** Was Kicked By:** {message.from_user.mention if message.from_user else 'Anon'}
-**Reason:** {reason or 'No Reason Provided.'}"""
+    msg = f"{mention}** Was Kicked By:** {message.from_user.mention if message.from_user else 'Anon'}\n**Reason:** {reason or 'No Reason Provided.'}"
     if message.command[0][0] == "d":
         await message.reply_to_message.delete()
     await message.chat.ban_member(user_id)
@@ -95,34 +74,25 @@ async def kickFunc(client, message: Message, _):
     await asyncio.sleep(1)
     await message.chat.unban_member(user_id)
 
-
-@app.on_message(command(BAN)& restrict_filter)
+@app.on_message(command("ban")& restrict_filter)
 @language
 async def banFunc(client, message: Message, _):
     user_id, reason = await extract_user_and_reason(message, sender_chat=True)
+    keyboard = InlineKeyboardMarkup([[InlineKeyboardButton(text="❗️Un ban",callback_data=f"_unban_{user_id}")]])
     if not user_id:
         return await message.reply_text(_["ban3"])
     if user_id == BOT_ID:
         return await message.reply_text(_["ban4"])
     st = await client.get_chat_member(message.chat.id, user_id)
-    if (
-                st.status == "administrator"
-                and st.status == "creator"
-        ):
-        return await message.reply_text("I can't ban admin in this group !")
+    if (st.status == "administrator" and st.status == "creator"):
+        return await app.send_message(chat_id = message.chat.id,text = "I can't ban admin in this group !")
     try:
         mention = (await app.get_users(user_id)).mention
     except IndexError:
-        mention = (
-            message.reply_to_message.sender_chat.title
-            if message.reply_to_message
-            else "Anon"
-        )
+        mention = (message.reply_to_message.sender_chat.title if message.reply_to_message else "Anon")
     await message.chat.ban_member(user_id)
-    msg = (
-        f"{mention}\n"
-        f"**Was Banned By:** {message.from_user.mention if message.from_user else 'Anon'}\n"
-    )
+    msg = f"{mention} **Was Banned By:** {message.from_user.mention if message.from_user else 'Anon'}\n"
+    await message.reply_text(msg,reply_markup=keyboard)
     if message.command[0][0] == "d":
         await message.reply_to_message.delete()
     if message.command[0] == "tban":
@@ -145,21 +115,16 @@ async def banFunc(client, message: Message, _):
     if reason:
         msg += f"**Reason:** {reason}"
     await message.chat.ban_member(user_id)
-    await message.reply_text(msg)
+    await message.reply_text(msg,reply_markup=keyboard)
 
-
-@app.on_message(command(UNBAN)  & restrict_filter)
+@app.on_message(command("unban")  & restrict_filter)
 @language
 async def unbanFunc(client, message: Message, _):
     if len(message.text.split()) == 1 and not message.reply_to_message:
         await message.reply_text("Provide a username or reply to a user's message to unban.")
         await message.stop_propagation()
-
     if message.reply_to_message and not message.reply_to_message.from_user:
-        user_id, user_first_name = (
-            message.reply_to_message.sender_chat.id,
-            message.reply_to_message.sender_chat.title,
-        )
+        user_id, user_first_name = (message.reply_to_message.sender_chat.id,message.reply_to_message.sender_chat.title)
     else:
         try:
             user_id, user_first_name, _ = await extract_user(app, message)
@@ -168,8 +133,7 @@ async def unbanFunc(client, message: Message, _):
     await message.chat.unban_member(user_id)
     await message.reply_text(f"{user_first_name} was unbanned by {message.from_user.mention}")
 
-
-@app.on_message(command(SBAN) & restrict_filter)
+@app.on_message(command("sban") & restrict_filter)
 async def kickunc(client, message: Message, _):
     if len(message.text.split()) == 1 and not message.reply_to_message:
         return
@@ -178,13 +142,9 @@ async def kickunc(client, message: Message, _):
     except Exception:
         return   
     if not user_id:
-        await message.reply_text("Can't find user to kick")
-        return   
+        return await message.reply_text("Can't find user to ban")   
     st = await app.get_chat_member(message.chat.id, user_id)
-    if (
-                st.status == "administrator"
-                and st.status == "creator"
-        ):
+    if (st.status == "administrator" and st.status == "creator"):
         return     
     try:
         await message.chat.ban_member(user_id)
@@ -192,11 +152,11 @@ async def kickunc(client, message: Message, _):
         if message.reply_to_message:
             await message.reply_to_message.delete()
     except Exception as ef:
-        await message.reply_text(f"{ef}")
+        await message.reply_text(ef)
     return
 
 
-__MODULE__ = f"{Restrict}"
+__MODULE__ = Restrict
 __HELP__ = """
 Some people need to be publicly banned; spammers, annoyances, or just trolls.
